Fix frozen remote players: epoch-scale ticks vs float interpolation time
Remote puppets never moved on Linux-hosted servers. Server::GetTick()
returned high_resolution_clock time_since_epoch() in ms — on Linux that
clock is system_clock, so ticks were ~1.75e12. InterpolateEntity cast
tick time to float, whose 24-bit mantissa snaps values of that magnitude
to ~131-second steps: 'tick > future.Tick' stayed false for minutes, the
timepoint queue never popped, and every remote puppet stayed pinned to
its first received position (verified live: hopping positions arrived
and were queued while SetTransform applied a constant for minutes).
Upstream never saw this because Windows' high_resolution_clock is
QPC-since-boot — small values float can represent to the millisecond.

- Server: ticks are now server-start-relative (small by construction).
- Client: interpolation time math float -> double (exact to 2^53 ms),
  guarding against any other server's tick origin.
- MOVEDIAG: add current-vs-target probe in GetDeltaTransform so the
  verification run proves actual in-world displacement from logs.

+++ code/client/App/World/InterpolationSystem.cpp
@@ -37,7 +37,13 @@ inline void SetSimpleMovement(Red::vehicle::IMoveSystem* apMoveSystem, const Red
 
 void InterpolateEntity(flecs::entity aEntity, const EntityComponent& aEntityComponent, InterpolationComponent& aInterpolation, float aSimulationDelay, Red::vehicle::IMoveSystem* apMoveSystem)
 {
-    const float tick = NetworkWorldSystem::GetTick() - aSimulationDelay;
+    // double, NOT float: ticks are absolute milliseconds. A float mantissa
+    // (24 bits) snaps epoch-scale ticks (~1.75e12 from a unix-clock server) to
+    // ~131-second steps, so `tick > future.Tick` stays false for minutes and
+    // every remote puppet freezes on its first received position. double is
+    // exact to 2^53 ms. (The server now also sends small start-relative ticks;
+    // this guards against any other server implementation.)
+    const double tick = static_cast<double>(NetworkWorldSystem::GetTick()) - aSimulationDelay;
 
     while (!aInterpolation.TimePoints.empty())
     {
